@@ -8,9 +8,9 @@ namespace BlackSlope.Api.Common.Versioning.Services
     public class JsonVersionService : IVersionService
     {
         private readonly IFileSystem _fileSystem;
-        private readonly IHostingEnvironment _hostingEnvironment;
+        private readonly IWebHostEnvironment _hostingEnvironment;
 
-        public JsonVersionService(IFileSystem fileSystem, IHostingEnvironment hostingEnvironment)
+        public JsonVersionService(IFileSystem fileSystem, IWebHostEnvironment hostingEnvironment)
         {
             _fileSystem = fileSystem;
             _hostingEnvironment = hostingEnvironment;
@@ -18,7 +18,7 @@ namespace BlackSlope.Api.Common.Versioning.Services
 
         public Version GetVersion()
         {
-            string filepath = _fileSystem.Path.Combine(_hostingEnvironment.ContentRootPath, "Common", "Version", "version.json");
+            var filepath = _fileSystem.Path.Combine(_hostingEnvironment.ContentRootPath, "Common", "Version", "version.json");
             var fileContents = _fileSystem.File.ReadAllText(filepath);
             dynamic task = JObject.Parse(fileContents);
             return new Version(task.version.ToString());
