@@ -1,5 +1,7 @@
-﻿using BlackSlope.Api.Common.Configuration;
+﻿using System.Collections.Generic;
+using BlackSlope.Api.Common.Configuration;
 using Microsoft.AspNetCore.Builder;
+using Microsoft.OpenApi.Models;
 
 namespace BlackSlope.Api.Common.Extensions
 {
@@ -9,8 +11,10 @@ namespace BlackSlope.Api.Common.Extensions
         {
             app.UseSwagger(c =>
             {
-                c.PreSerializeFilters.Add((swagger, httpReq) => swagger.Host = httpReq.Host.Value);
+                c.SerializeAsV2 = true;
+                c.PreSerializeFilters.Add((swagger, httpReq) => swagger.Servers = new List<OpenApiServer> { new OpenApiServer { Url = $"{httpReq.Scheme}://{httpReq.Host.Value}" } });
             });
+
             return app.UseSwaggerUI(c =>
             {
                 c.SwaggerEndpoint("/swagger/" + swaggerConfig.Version + "/swagger.json", swaggerConfig.ApplicationName);
