@@ -37,7 +37,7 @@ namespace BlackSlope.Hosts.ConsoleApp
         public static async Task<AuthenticationResult> GetTokenAsynch(string clientId, string key, string tenantId, string appIdUniformResourceIdentifier)
         {
             var aadInstanceUrl = "https://login.microsoftonline.com/{0}";
-            var authority = String.Format(CultureInfo.InvariantCulture, aadInstanceUrl, tenantId);
+            var authority = string.Format(CultureInfo.InvariantCulture, aadInstanceUrl, tenantId);
 
             var authContext = new AuthenticationContext(authority);
             var clientCredential = new ClientCredential(clientId, key);
